requirements.txt
Flask==2.0.1
jsonpickle
numpy
pandas
Pillow
scipy
torch==1.9.0
tqdm
